Add iam instance profile support
In order to better use SSM for building rather than SSH, we need to be
able to configure the iam_instance_profile.  This change adds
`iam_instance_profile` that can be used when setting `ssh_interfance` to
`session_manager`.  When using `session_manager` you will need to use a
non-minimal version of the source ami to ensure it has the SSM agent
available at launch. You may also need to increase the volume size from
4 depending on which source AMI you use.